Move entities from state to src folder
