+++ packages/game/src/entities/Player.ts
@@ -1,13 +1,18 @@
 import { Schema, type } from '@colyseus/schema'
-import { Transform, PlayerData } from '../../components'
+import { annotations } from '@pixatore/ecs'
+import { Transform, PlayerData } from '../components'
+import { ComponentTypes } from '../components/ComponentTypes'
 
+@annotations.ecsEntity
 export class Player extends Schema {
   @type('string')
   public id: string
 
+  @annotations.ecsComponent(ComponentTypes.Transform)
   @type(Transform)
   public transform: Transform
 
+  @annotations.ecsComponent(ComponentTypes.PlayerData)
   @type(PlayerData)
   public data: PlayerData
