+++ packages/game/src/state/GameState.ts
@@ -1,5 +1,5 @@
 import { MapSchema, Schema, type } from '@colyseus/schema'
-import { Player } from './entities/Player'
+import { Player } from '../entities/Player'
 import { Status } from '../components/Status'
 
 export class GameState extends Schema {
